Implement users permissions

+++ app/controllers/users_controller.rb
@@ -1,4 +1,7 @@
 class UsersController < ApplicationController
+  before_action :set_user, only: [:show, :edit, :update, :destroy]
+  before_action :authenticate_user!, except: [:new]
+
   def new
     @user = User.new
   end
@@ -14,12 +17,13 @@ class UsersController < ApplicationController
     end
   end
 
-  def index
-    @users = User.all
+  def edit
+      redirect_to :root
+      flash[:alert] = "Edit your profile by clicking your 'profile' link"
   end
 
-  def edit
-    @user = current_user
+  def index
+    @users = User.all
   end
 
   def show
@@ -27,14 +31,18 @@ class UsersController < ApplicationController
   end
 
   def destroy
-    reviews = Review.where(user_id: current_user)
-    reviews.destroy_all
-    foods = Food.where(user_id: current_user)
-    foods.destroy_all
-    current_user.destroy
-    respond_to do |format|
-      format.html { redirect_to root_path, notice: 'Profile was successfully delete.' }
-      format.json { head :no_content }
+    unless @user.destroyable_by?(current_user, @user)
+      redirect_to :root
+    else
+      reviews = Review.where(user_id: current_user)
+      reviews.destroy_all
+      foods = Food.where(user_id: current_user)
+      foods.destroy_all
+      current_user.destroy
+      respond_to do |format|
+        format.html { redirect_to root_path, notice: 'Profile was successfully delete.' }
+        format.json { head :no_content }
+      end
     end
   end
